requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
billiard==4.2.0
celery==5.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.7
django-celery-results==2.5.1
django-filter==23.3
django-jalali==6.0.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
executing==2.0.1
flower==2.0.1
humanize==4.8.0
inflection==0.5.1
ipython==8.17.2
jdatetime==4.1.1
jedi==0.19.1
kombu==5.3.4
Markdown==3.5.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
Pillow==10.1.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.10
# uWSGI==2.0.23